+++ R/add_to_quarto.R
@@ -7,44 +7,34 @@
 #' @param quarto_dir The base directory for your quarto project
 #' @param include_dir The directory where you want to put the css and
 #'   js files (defaults to "include")
-#' @param script_dir The directory where you want to put the .R script
-#'   (defaults to "R")
 #' @param output_format The format you want to add
 #'   webexercises to (only html for now)
-#' @param render Whether to render the book after updating (defaults to FALSE).
 #'
 #' @return No return value, called for side effects.
 #' @export
 #'
 add_to_quarto <- function(quarto_dir = ".",
                           include_dir = "include",
-                          script_dir = "R",
-                          output_format = c("html"),
-                          render = FALSE) {
+                          output_format = c("html")) {
   # check inputs
   if (quarto_dir == "") bookdown_dir <- "."
   if (include_dir == "") include_dir <- "."
-  if (script_dir == "") script_dir <- "."
   output_format <- match.arg(output_format)
 
   # get helper files
   css <- system.file("reports/default/webex.css", package = "webexercises")
   js <- system.file("reports/default/webex.js", package = "webexercises")
-  script <- system.file("reports/default/webex.R", package = "webexercises")
-  index <- system.file("reports/default/index.qmd", package = "webexercises")
+  demo <- system.file("reports/default/webexercises.qmd", package = "webexercises")
 
   # make sure include and script directories exist
   incdir <- file.path(quarto_dir, include_dir)
   dir.create(path = incdir, showWarnings = FALSE, recursive = TRUE)
 
-  rdir <- file.path(quarto_dir, script_dir)
-  dir.create(path = rdir, showWarnings = FALSE, recursive = TRUE)
-
   # add or update helper files
   file.copy(css, incdir, overwrite = TRUE)
   file.copy(js, incdir, overwrite = TRUE)
-  file.copy(script, rdir, overwrite = TRUE)
-  message("webex.css, webex.js, and webex.R updated")
+  file.copy(demo, quarto_dir, overwrite = TRUE)
+  message("webex.css, webex.js, and webexercises.qmd updated")
 
   # update or create _quarto.yml
   css_path <- file.path(include_dir, "webex.css")
@@ -86,47 +76,13 @@ add_to_quarto <- function(quarto_dir = ".",
   yaml::write_yaml(yml, quarto_file)
   message(quarto_file, " updated")
 
-  # add setup script to .Rprofile
+  # update .Rprofile
   rprofile <- file.path(quarto_dir, ".Rprofile")
-  source <- paste0('source("', script_dir, '/webex.R")')
-  write(source, rprofile, append = TRUE)
 
-  # add webexercises file
-  webex_path <- file.path(quarto_dir, "webexercises.qmd")
-  file.copy(index, webex_path, overwrite = FALSE)
-  webex_file <- readLines(webex_path)
-  webex_file <- sapply(webex_file, gsub,
-                       pattern = "webex.R",
-                       replacement = file.path(script_dir, "webex.R"),
-                       fixed = TRUE, USE.NAMES = FALSE)
-  webex_file <- sapply(webex_file, gsub,
-                       pattern = "webex.css",
-                       replacement = css_path,
-                       fixed = TRUE, USE.NAMES = FALSE)
-  webex_file <- sapply(webex_file, gsub,
-                       pattern = "webex.js",
-                       replacement = js_path,
-                       fixed = TRUE, USE.NAMES = FALSE)
-  write(webex_file, webex_path)
-
-  # render and open site
-  if (isTRUE(render)) {
-    if (!requireNamespace("quarto", quietly = TRUE)) {
-      stop("Package \"quarto\" needed for this function to work. Please install it.",
-           call. = FALSE)
-    }
-    if (!requireNamespace("xfun", quietly = TRUE)) {
-      stop("Package \"xfun\" needed for this function to work. Please install it.",
-           call. = FALSE)
-    }
-
-    xfun::in_dir(quarto_dir, quarto::quarto_render(
-      input = "webexercises.qmd",
-      output_format = output_format
-    ))
-
-    utils::browseURL(file.path(quarto_dir, "webexercises.html"))
-  }
+  load_txt <- "# load webexercises before each chapter
+# needs to check namespace to not bork github actions
+if (requireNamespace('webexercises', quietly = TRUE)) library(webexercises)"
+  write(load_txt, rprofile, append = TRUE)
 
   invisible(NULL)
 }