+++ tests/testthat/test-add_to_quarto.R
@@ -7,18 +7,16 @@ test_that("defaults", {
 
   write("project:\n  title: quarto_test\n", "quarto_test/_quarto.yml")
 
-  add_to_quarto("quarto_test", render = FALSE)
+  add_to_quarto("quarto_test")
 
   quarto <- file.path("quarto_test", "_quarto.yml")
   css <- file.path("quarto_test", "include", "webex.css")
   js <- file.path("quarto_test", "include", "webex.js")
-  script <- file.path("quarto_test", "R", "webex.R")
   rprofile <- file.path("quarto_test", ".Rprofile")
   webex <- file.path("quarto_test", "webexercises.qmd")
 
   expect_true(file.exists(css))
   expect_true(file.exists(js))
-  expect_true(file.exists(script))
   expect_true(file.exists(rprofile))
   expect_true(file.exists(webex))
 
@@ -35,20 +33,16 @@ test_that("names dirs", {
   write("project:\n  title: quarto_test\n", "quarto_test/_quarto.yml")
 
   add_to_quarto(quarto_dir = "quarto_test",
-                include_dir = "",
-                script_dir = "",
-                render = FALSE)
+                include_dir = "")
 
   quarto <- file.path("quarto_test", "_quarto.yml")
   css <- file.path("quarto_test", "webex.css")
   js <- file.path("quarto_test", "webex.js")
-  script <- file.path("quarto_test", "webex.R")
   rprofile <- file.path("quarto_test", ".Rprofile")
   webex <- file.path("quarto_test", "webexercises.qmd")
 
   expect_true(file.exists(css))
   expect_true(file.exists(js))
-  expect_true(file.exists(script))
   expect_true(file.exists(rprofile))
   expect_true(file.exists(webex))